refactor(customer-portal): Restructure authentication context and hooks
