+++ apps/customer-portal/src/pages/Dashboard.jsx
@@ -1,18 +1,16 @@
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useAuth } from '../contexts/AuthContext';
+import useAuth from '../hooks/useAuth';
 import {
     Container,
     Typography,
     Box,
     Paper,
-    Button,
     AppBar,
     Toolbar,
     IconButton,
     Menu,
-    MenuItem,
-    Avatar
+    MenuItem
 } from '@mui/material';
 import { AccountCircle } from '@mui/icons-material';
